pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.3.3", "pybind11", "setuptools-scm>=8.0", "cmake<3.30 ; platform_system=='Darwin'"]
build-backend = "scikit_build_core.build"


[project]
name = "sasktran2"
dynamic = ["version"]
dependencies = ['numpy', 'appdirs', 'pyyaml', 'xarray>=2023.0.0', 'scipy', 'netcdf4']
license = {text = "MIT"}
description= "The SASKTRAN radiative transfer model Version 2"
readme = "README.md"
authors = [
    { name = "Daniel Zawada", email = "daniel.zawada@usask.ca" },
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
test = ["pytest"]
solar = ["astropy"]
hapi = ["hitran-api"]
dev = ["ruff", "sphinx", "sphinx_book_theme", "myst-nb", "sphinx_design"]
complete = ["sasktran2[solar,hapi]"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64", "linux-aarch64", "linux-ppc64le", "osx-64", "osx-arm64", "win-64"]

[tool.pixi.pypi-dependencies]

[tool.pixi.pypi-options]

[tool.pixi.tasks]
build = { cmd = [
    "pip",
    "install",
    "--no-build-isolation",
    "-v",
    "-e",
    ".",
] }
test = { cmd = "pytest tests/", cwd = "." }
pre-commit = { cmd = "pre-commit run -a", cwd = "." }
docs = { cmd = "sphinx-build -b html docs/sphinx/source docs/sphinx/build", cwd = "." }
update-stubs = { cmd = "pybind11-stubgen sasktran2._core -o src/", cwd = "." }


[tool.pixi.dependencies]
numpy = "*"
appdirs = "*"
pyyaml = "*"
xarray = ">=2023.0.0"
scipy = "*"
netcdf4 = "*"
pytest = ">=8.3.3,<9"
astropy = "*"
ruff = "*"
sphinx = "*"
myst-nb = "*"
openblas = "*"
libopenblas = "*"
eigen = "*"
spdlog = "*"
c-compiler = "*"
cxx-compiler = "*"
fortran-compiler = "*"
scikit-build-core = "*"
cmake = "*"
pip = "*"
setuptools_scm = "*"
pybind11 = "*"
matplotlib = "*"
pybind11-stubgen = "*"
pre-commit = "*"
seaborn = ">=0.13.2,<0.14"

[tool.pixi.environments]
default = ["test", "solar", "hapi", "dev"]


[tool.scikit-build]
wheel.expand-macos-universal-tags = true
metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
sdist.include = ["src/sasktran2/_version.py"]

[tool.scikit-build.cmake.define]
DO_STREAM_TEMPLATES = {env="DO_STREAM_TEMPLATES", default="OFF"}

[tool.setuptools_scm]
write_to = "src/sasktran2/_version.py"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
    "ignore:(ast.Str|Attribute s|ast.NameConstant|ast.Num) is deprecated:DeprecationWarning:_pytest",  # Python 3.12
]
testpaths = ["tests"]


[tool.cibuildwheel]
test-command = "bash {project}/ci/cibw/cibw_test_script.sh {project}"
test-extras = ["test"]
test-skip = ["*universal2:arm64"]
# Setuptools bug causes collision between pypy and cpython artifacts
# before-build = "rm -rf {project}/build"
before-build = "bash {project}/ci/cibw/cibw_before_build.sh {project}"

[tool.cibuildwheel.linux]
archs = "native"
environment = { CFLAGS="-fno-strict-aliasing -DBLAS_SYMBOL_SUFFIX=64_ -DHAVE_BLAS_ILP64", LDFLAGS="-Wl,--strip-debug", CXXFLAGS="-DBLAS_SYMBOL_SUFFIX=64_ -DHAVE_BLAS_ILP64", RUNNER_OS="Linux", CMAKE_BUILD_PARALLEL_LEVEL="2", DO_STREAM_TEMPLATES="ON"}

[tool.cibuildwheel.macos]
archs = "native"
test-skip = "*_universal2:arm64"
# MACOS linker doesn't support stripping symbols
environment = {CFLAGS="-fno-strict-aliasing -DBLAS_SYMBOL_SUFFIX=64_ -DHAVE_BLAS_ILP64", CXXFLAGS="-DBLAS_SYMBOL_SUFFIX=64_ -DHAVE_BLAS_ILP64",  CC="clang", CXX = "clang++", RUNNER_OS="macOS", CMAKE_ARGS="-DUSE_OPENMP=OFF", CMAKE_BUILD_PARALLEL_LEVEL="3", DO_STREAM_TEMPLATES="ON"}

[tool.cibuildwheel.windows]
archs = "native"
environment = {CFLAGS="-DBLAS_SYMBOL_SUFFIX=64_ -DHAVE_BLAS_ILP64", CXXFLAGS="-DBLAS_SYMBOL_SUFFIX=64_ -DHAVE_BLAS_ILP64", LDFLAGS="", PKG_CONFIG_PATH="C:/opt/64/lib/pkgconfig", CMAKE_PREFIX_PATH="C:/opt/64/", CMAKE_BUILD_PARALLEL_LEVEL="2", DO_STREAM_TEMPLATES="ON"}
#config-settings = "setup-args=--vsenv"
repair-wheel-command = "bash ./ci/cibw/repair_windows.sh {wheel} {dest_dir}"


[tool.ruff]
src = ["python"]
exclude = ["tools/*", "docs/sphinx/source/conf.py", "src/sasktran2/_core.pyi"]

[tool.ruff.lint]
select = [
    "E", "F", "W", # flake8
    "B",           # flake8-bugbear
    "I",           # isort
    "ARG",         # flake8-unused-arguments
    "C4",          # flake8-comprehensions
    "EM",          # flake8-errmsg
    "ICN",         # flake8-import-conventions
    "ISC",         # flake8-implicit-str-concat
    "G",           # flake8-logging-format
    "PGH",         # pygrep-hooks
    "PIE",         # flake8-pie
    "PL",          # pylint
    "PT",          # flake8-pytest-style
    "PTH",         # flake8-use-pathlib
    "RET",         # flake8-return
    "RUF",         # Ruff-specific
    "SIM",         # flake8-simplify
    "T20",         # flake8-print
    "UP",          # pyupgrade
    "YTT",         # flake8-2020
    "EXE",         # flake8-executable
    "NPY",         # NumPy specific rules
    "PD",          # pandas-vet
]
extend-ignore = [
    "PLR",    # Design related pylint codes
    "E501",   # Line too long
    "PD011", # wanting to use to_numpy instead of values
    "G004",   # Logging format
]
unfixable = [
    "T20",  # Removes print statements
    "F841", # Removes unused variables
    "F401", # Removes unused imports
]
flake8-unused-arguments.ignore-variadic-names = true
isort.required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
